refactor: split high-impact files, reduce complexity per fallow findings
