+++ src/test/extension/hexScopeStorage.test.ts
@@ -26,7 +26,8 @@ import {
 } from '../../hexScopeStorage';
 import type { MementoLike } from '../../hexScopeMigration';
 import { migrateLegacyData } from '../../hexScopeMigration';
-import { migrateStructDefinitions, normalizeStructDefsValue } from '../../core/structMigration';
+import { migrateStructDefinitions } from '../../core/structMigration';
+import { normalizeStructDefsValue } from '../../core/structNormalization';
 
 const sleep = (ms: number) => new Promise<void>(resolve => setTimeout(resolve, ms));
 const FAST = 1; // near-immediate debounce for tests
