+++ src/webview/editTransactions.ts
@@ -1,60 +1,5 @@
-import type { SelectionRange } from './memory/selection';
 import { S } from './state';
-
-export function stageIntegrityEditTransaction(edits: Array<[number, number]>): number {
-    const previous: Array<[number, number]> = [];
-    for (const [address, value] of edits) {
-        const prior = stageIntegrityEdit(address, value);
-        if (prior) { previous.push(prior); }
-    }
-    if (previous.length === 0) { return 0; }
-    S.undoStack.push(previous);
-    S.redoStack.length = 0;
-    S.editMode = true;
-    return previous.length;
-}
-
-export function stageIntegrityEdit(address: number, value: number): [number, number] | null {
-    const original = getOriginalByte(address);
-    if (original === undefined) { return null; }
-    const current = currentIntegrityByte(address, original);
-    if (current === value) { return null; }
-    if (value === original) { S.edits.delete(address); }
-    else { S.edits.set(address, value); }
-    return [address, current];
-}
-
-function currentIntegrityByte(address: number, original: number): number {
-    return S.edits.has(address) ? S.edits.get(address)! : original;
-}
-
-export function fillSelectionTransaction(range: SelectionRange | null, fillVal: number): void {
-    const prev = buildFillTransaction(range, fillVal);
-    if (prev.length > 0) {
-        S.undoStack.push(prev);
-        S.redoStack.length = 0;
-    }
-}
-
-function buildFillTransaction(range: SelectionRange | null, fillVal: number): Array<[number, number]> {
-    if (!range) { return []; }
-    const prev: Array<[number, number]> = [];
-    for (let a = range.start; a <= range.end; a++) {
-        const prior = stageFillByte(a, fillVal);
-        if (prior) { prev.push(prior); }
-    }
-    return prev;
-}
-
-function stageFillByte(address: number, fillVal: number): [number, number] | null {
-    const original = getOriginalByte(address);
-    if (original === undefined) { return null; }
-    const current = currentIntegrityByte(address, original);
-    if (current === fillVal) { return null; }
-    if (fillVal === original) { S.edits.delete(address); }
-    else { S.edits.set(address, fillVal); }
-    return [address, current];
-}
+import { currentEditedByte, restoreEditedByte } from './editByteState';
 
 export function undoLastEditTransaction(): boolean {
     const txn = popUndoTransaction();
@@ -90,37 +35,4 @@ function popRedoTransaction(): Array<[number, number]> | null {
     if (!S.editMode) { return null; }
     if (S.redoStack.length === 0) { return null; }
     return S.redoStack.pop()!;
-}
-
-/** Current in-effect byte value at `addr` (edited value or the original). */
-function currentEditedByte(addr: number): number {
-    const orig = getOriginalByte(addr);
-    return S.edits.has(addr) ? S.edits.get(addr)! : (orig ?? 0);
-}
-
-function restoreEditedByte(addr: number, prevVal: number): void {
-    const orig = getOriginalByte(addr);
-    if (orig !== undefined && prevVal === orig) {
-        S.edits.delete(addr);
-        return;
-    }
-    S.edits.set(addr, prevVal);
-}
-
-/** Mass-revert staged edits to snapshot values (selection-session discard). */
-export function restoreEditedBytes(prev: Array<[number, number]>): void {
-    for (const [addr, prevVal] of prev) { restoreEditedByte(addr, prevVal); }
-}
-
-function getOriginalByte(addr: number): number | undefined {
-    if (!S.parseResult) { return undefined; }
-    for (const seg of S.parseResult.segments) {
-        const off = addr - seg.startAddress;
-        if (isSegmentOffset(off, seg.data.length)) { return seg.data[off]; }
-    }
-    return undefined;
-}
-
-function isSegmentOffset(offset: number, length: number): boolean {
-    return offset >= 0 && offset < length;
 }
